exploit now works
1. automatically generates payload and "tricks" admin to log in on cross site scripted login page

2. places evil php script in cms that allows reverse shell execution

3. reverse shell waits for stdin command and executes it on victim
