+++ scenarios/CVE-2020-23839/main.py
@@ -53,6 +53,12 @@ if __name__ == '__main__':
 
     storage_services = [JSONFileStorage()]
 
+    # Attack steps
+
+    trick_admin = ExecCommand("")
+
+    execute_reverse_shell = ExecCommand("")
+
     victim = Image("victim_xss")
     normal = Image("normal_xss", command=StdinCommand(""), init_args="-ip ${victim}")
     exploit = Image("exploit_xss", command=ExecCommand("python3 /home/exploit.py -ip ${victim}"))
